fixing to only take n-1 kerfs per cut group
